+++ src/zfw.c
@@ -230,7 +230,6 @@ struct bpf_event
     __u8 tracking_code;
     unsigned char source[6];
     unsigned char dest[6];
-    char service_id[29];
 };
 
 struct diag_ip4
@@ -1905,14 +1904,13 @@ static int process_events(void *ctx, void *data, size_t len)
     char *ts = get_ts(evt->tstamp);
     char message[250];
     int res = 0;
-    char *service_id = evt->service_id;
     if (((ifname && monitor_interface && !strcmp(monitor_interface, ifname)) || all_interface) && ts)
     {
         if (evt->error_code)
         {
             if (evt->error_code == IP_HEADER_TOO_BIG)
             {
-                sprintf(message, "%s : %-22s : %s : %s : IP Header Too Big\n", ts, service_id, ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
+                sprintf(message, "%s : %s : %s : IP Header Too Big\n", ts, ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
                 if (logging)
                 {
                     res = write_log(log_file_name, message);
@@ -1924,7 +1922,7 @@ static int process_events(void *ctx, void *data, size_t len)
             }
             else if (evt->error_code == NO_IP_OPTIONS_ALLOWED)
             {
-                sprintf(message, "%s : %-22s : %s : %s : No IP Options Allowed\n", ts, service_id, ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
+                sprintf(message, "%s : %s : %s : No IP Options Allowed\n", ts, ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
                 if (logging)
                 {
                     res = write_log(log_file_name, message);
@@ -1936,7 +1934,7 @@ static int process_events(void *ctx, void *data, size_t len)
             }
             else if (evt->error_code == UDP_HEADER_TOO_BIG)
             {
-                sprintf(message, "%s : %-22s : %s : %s : UDP Header Too Big\n", ts, service_id, ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
+                sprintf(message, "%s : %s : %s : UDP Header Too Big\n", ts, ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
                 if (logging)
                 {
                     res = write_log(log_file_name, message);
@@ -1948,7 +1946,7 @@ static int process_events(void *ctx, void *data, size_t len)
             }
             else if (evt->error_code == GENEVE_HEADER_TOO_BIG)
             {
-                sprintf(message, "%s : %-22s : %s : %s : Geneve Header Too Big\n", ts, service_id ,ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
+                sprintf(message, "%s : %s : %s : Geneve Header Too Big\n", ts, ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
                 if (logging)
                 {
                     res = write_log(log_file_name, message);
@@ -1960,7 +1958,7 @@ static int process_events(void *ctx, void *data, size_t len)
             }
             else if (evt->error_code == GENEVE_HEADER_LENGTH_VERSION_ERROR)
             {
-                sprintf(message, "%s : %-22s : %s : %s : Geneve Header Length: Version Error\n", ts, service_id ,ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
+                sprintf(message, "%s : %s : %s : Geneve Header Length: Version Error\n", ts, ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
                 if (logging)
                 {
                     res = write_log(log_file_name, message);
@@ -1972,7 +1970,7 @@ static int process_events(void *ctx, void *data, size_t len)
             }
             else if (evt->error_code == SKB_ADJUST_ERROR)
             {
-                sprintf(message, "%s : %-22s : %s : %s : SKB Adjust Error\n", ts, service_id ,ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
+                sprintf(message, "%s : %s : %s : SKB Adjust Error\n", ts, ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
                 if (logging)
                 {
                     res = write_log(log_file_name, message);
@@ -1984,7 +1982,7 @@ static int process_events(void *ctx, void *data, size_t len)
             }
             else if (evt->error_code == ICMP_HEADER_TOO_BIG)
             {
-                sprintf(message, "%s : %-22s : %s : %s : ICMP Header Too Big\n", ts, service_id , ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
+                sprintf(message, "%s : %s : %s : ICMP Header Too Big\n", ts, ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
                 if (logging)
                 {
                     res = write_log(log_file_name, message);
@@ -1996,7 +1994,7 @@ static int process_events(void *ctx, void *data, size_t len)
             }
             else if (evt->error_code == ICMP_INNER_IP_HEADER_TOO_BIG)
             {
-                sprintf(message, "%s : %-22s : %s : %s : ICMP Inner IP Header Too Big\n", ts, service_id,ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
+                sprintf(message, "%s : %s : %s : ICMP Inner IP Header Too Big\n", ts, ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
                 if (logging)
                 {
                     res = write_log(log_file_name, message);
@@ -2008,7 +2006,7 @@ static int process_events(void *ctx, void *data, size_t len)
             }
             else if (evt->error_code == IF_LIST_MATCH_ERROR)
             {
-                sprintf(message, "%s : %-22s : %s : %s : Interface did not match and per interface filtering is enabled\n", ts, service_id ,ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
+                sprintf(message, "%s : %s : %s : Interface did not match and per interface filtering is enabled\n", ts, ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
                 if (logging)
                 {
                     res = write_log(log_file_name, message);
@@ -2020,7 +2018,7 @@ static int process_events(void *ctx, void *data, size_t len)
             }
             else if (evt->error_code == NO_REDIRECT_STATE_FOUND)
             {
-                sprintf(message, "%s : %-22s : %s : %s : No Redirect State found\n", ts, service_id, ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
+                sprintf(message, "%s : %s : %s : No Redirect State found\n", ts, ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS");
                 if (logging)
                 {
                     res = write_log(log_file_name, message);
@@ -2054,7 +2052,7 @@ static int process_events(void *ctx, void *data, size_t len)
                 char *tun_ifname = if_indextoname(evt->tun_ifindex, tbuf);
                 if (tun_ifname)
                 {
-                    sprintf(message, "%s : %-22s: %s : %s : %s:%d[%x:%x:%x:%x:%x:%x] > %s:%d[%x:%x:%x:%x:%x:%x] redirect ---> %s\n", ts, service_id, ifname, protocol, saddr, ntohs(evt->sport),
+                    sprintf(message, "%s : %s : %s :%s:%d[%x:%x:%x:%x:%x:%x] > %s:%d[%x:%x:%x:%x:%x:%x] redirect ---> %s\n", ts, ifname, protocol, saddr, ntohs(evt->sport),
                             evt->source[0], evt->source[1], evt->source[2], evt->source[3], evt->source[4], evt->source[5], daddr, ntohs(evt->dport),
                             evt->dest[0], evt->dest[1], evt->dest[2], evt->dest[3], evt->dest[4], evt->dest[5], tun_ifname);
                     if (logging)
@@ -2069,8 +2067,8 @@ static int process_events(void *ctx, void *data, size_t len)
             }
             else if (evt->tport && ifname)
             {
-                sprintf(message, "%s : %-22s : %s : %s : %s : %s:%d > %s:%d | tproxy ---> 127.0.0.1:%d\n",
-                        ts, service_id, ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS", protocol, saddr, ntohs(evt->sport),
+                sprintf(message, "%s : %s : %s : %s :%s:%d > %s:%d | tproxy ---> 127.0.0.1:%d\n",
+                        ts, ifname, (evt->direction == INGRESS) ? "INGRESS" : "EGRESS", protocol, saddr, ntohs(evt->sport),
                         daddr, ntohs(evt->dport), ntohs(evt->tport));
                 if (logging)
                 {
@@ -2136,7 +2134,7 @@ static int process_events(void *ctx, void *data, size_t len)
                 }
                 if (state)
                 {
-                    sprintf(message, "%s : %-22s : %s : %s : %s : %s:%d > %s:%d outbound_tracking ---> %s\n", ts, service_id, ifname,
+                    sprintf(message, "%s : %s : %s : %s :%s:%d > %s:%d outbound_tracking ---> %s\n", ts, ifname,
                             (evt->direction == INGRESS) ? "INGRESS" : "EGRESS", protocol, saddr, ntohs(evt->sport), daddr, ntohs(evt->dport), state);
                     if (logging)
                     {
@@ -2156,7 +2154,7 @@ static int process_events(void *ctx, void *data, size_t len)
                 if (code == 4)
                 {
                     /*evt->sport is use repurposed store next hop mtu*/
-                    sprintf(message, "%s : %-22s : %s : %s : %s : %s --> reported next hop mtu:%d > FRAGMENTATION NEEDED IN PATH TO:%s:%d\n", ts, service_id ,ifname,
+                    sprintf(message, "%s : %s : %s : %s :%s --> reported next hop mtu:%d > FRAGMENTATION NEEDED IN PATH TO:%s:%d\n", ts, ifname,
                             (evt->direction == INGRESS) ? "INGRESS" : "EGRESS", protocol, saddr, ntohs(evt->sport), daddr, ntohs(evt->dport));
                     if (logging)
                     {
@@ -2198,7 +2196,7 @@ static int process_events(void *ctx, void *data, size_t len)
 
                     if (code_string)
                     {
-                        sprintf(message, "%s : %-22s : %s : %s : %s : %s --> REPORTED:%s > in PATH TO:%s:%s:%d OUTER-TTL:%d INNER-TTL:%d\n", ts, service_id,ifname,
+                        sprintf(message, "%s : %s : %s : %s :%s --> REPORTED:%s > in PATH TO:%s:%s:%d OUTER-TTL:%d INNER-TTL:%d\n", ts, ifname,
                                 (evt->direction == INGRESS) ? "INGRESS" : "EGRESS", protocol, saddr, code_string, daddr, protocol_string, ntohs(evt->dport), outer_ttl, inner_ttl);
                         if (logging)
                         {
@@ -2213,7 +2211,7 @@ static int process_events(void *ctx, void *data, size_t len)
             }
             else if (ifname)
             {
-                sprintf(message, "%s : %-22s : %s : %s : %s : %s:%d > %s:%d\n", ts, service_id,ifname,
+                sprintf(message, "%s : %s : %s : %s :%s:%d > %s:%d\n", ts, ifname,
                         (evt->direction == INGRESS) ? "INGRESS" : "EGRESS", protocol, saddr, ntohs(evt->sport), daddr, ntohs(evt->dport));
                 if (logging)
                 {